refactor(color): extract hex-to-rgba helpers into shared utility
Address PR review feedback by consolidating the duplicated color
conversion logic from TaskBoardViewComponent and DashboardHeaderComponent
into src/app/core/utils/color.utils.ts. The shared helper supports both
3-digit and 6-digit hex codes and lets callers opt into a default color.

Also use CYBERPUNK_COLORS.TODO as the header's default color instead of
hardcoding #e040fb, and wrap the two dashboard-header button class lines
that exceeded the 100-character style guide limit.
