+++ src/app/features/dashboard/components/dashboard-header.html
@@ -19,19 +19,20 @@
           <div class="flex items-center gap-3">
             <h1
               class="text-xl font-bold flex items-center gap-3 leading-none"
-              [style.color]="project.color || '#e040fb'"
+              [style.color]="project.color || defaultColor"
               [style.text-shadow]="'0 0 10px ' + getColorWithOpacity(project.color, 0.5)"
             >
               <span
                 class="w-3 h-3 rounded-sm shadow-[0_0_10px_currentColor]"
-                [style.background-color]="project.color || '#e040fb'"
-                [style.color]="project.color || '#e040fb'"
+                [style.background-color]="project.color || defaultColor"
+                [style.color]="project.color || defaultColor"
               ></span>
               {{ project.name }}
             </h1>
 
             <button
-              class="p-1 rounded-lg text-slate-500 hover:text-white hover:bg-white/5 transition-colors self-center flex items-center justify-center"
+              class="p-1 rounded-lg text-slate-500 hover:text-white hover:bg-white/5
+                     transition-colors self-center flex items-center justify-center"
               (click)="showFieldManager.emit()"
               title="Manage Custom Fields"
             >
@@ -53,7 +54,8 @@
 
             <!-- Project Actions -->
             <button
-              class="p-1 rounded-lg text-slate-500 hover:text-white hover:bg-white/5 transition-colors self-center flex items-center justify-center"
+              class="p-1 rounded-lg text-slate-500 hover:text-white hover:bg-white/5
+                     transition-colors self-center flex items-center justify-center"
               (click)="editProjectModal.emit(project)"
               title="Edit Project"
             >
@@ -203,7 +205,7 @@
           class="p-2 rounded-lg border transition-colors"
           [style.background-color]="getColorWithOpacity(project.color, 0.15)"
           [style.border-color]="getColorWithOpacity(project.color, 0.4)"
-          [style.color]="project.color || '#e040fb'"
+          [style.color]="project.color || defaultColor"
           [style.box-shadow]="'0 0 12px ' + getColorWithOpacity(project.color, 0.25)"
           title="Project Settings"
         >
