+++ src/app/features/dashboard/components/dashboard-header.ts
@@ -1,8 +1,9 @@
 import { Component, input, output, inject, ChangeDetectionStrategy } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { RouterLink } from '@angular/router';
-import { Project, TaskViewMode } from '../../../core/models/domain.model';
+import { Project, TaskViewMode, CYBERPUNK_COLORS } from '../../../core/models/domain.model';
 import { AuthService } from '../../../core/auth/auth.service';
+import { getColorWithOpacity } from '../../../core/utils/color.utils';
 
 @Component({
   selector: 'app-dashboard-header',
@@ -13,6 +14,7 @@ import { AuthService } from '../../../core/auth/auth.service';
 })
 export class DashboardHeaderComponent {
   readonly auth = inject(AuthService);
+  readonly defaultColor = CYBERPUNK_COLORS.TODO;
 
   currentProject = input<Project | null>(null);
   viewMode = input<TaskViewMode>('list');
@@ -26,17 +28,6 @@ export class DashboardHeaderComponent {
   toggleSidebar = output<void>();
 
   getColorWithOpacity(color: string | undefined, alpha: number): string {
-    const hex = color || '#e040fb';
-    const parsed = hex.replace('#', '');
-    if (parsed.length !== 6) {
-      return `rgba(224, 64, 251, ${alpha})`;
-    }
-    const r = parseInt(parsed.substring(0, 2), 16);
-    const g = parseInt(parsed.substring(2, 4), 16);
-    const b = parseInt(parsed.substring(4, 6), 16);
-    if (Number.isNaN(r) || Number.isNaN(g) || Number.isNaN(b)) {
-      return `rgba(224, 64, 251, ${alpha})`;
-    }
-    return `rgba(${r}, ${g}, ${b}, ${alpha})`;
+    return getColorWithOpacity(color, alpha);
   }
 }
